requirements.txt
kafka-python==2.3.0
py4j==0.10.9.9
pyspark==3.5.8
websocket-client==1.8.0
websockets==13.1
